backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.25.2
python-jose==3.3.0
pytest==7.4.3
httpx==0.25.2
openai==1.13.3
langchain==0.1.10
langchain-openai==0.0.8
python-dotenv==1.0.1
markdown==3.5.2
jinja2==3.1.3
pandas==2.1.4
openpyxl==3.1.2

# Email and Scheduling
schedule==1.2.0
sendgrid==6.11.0
